+++ src/duqtools/config/_schema_create.py
@@ -3,21 +3,25 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Optional, Union
 
-from pydantic import Field, PrivateAttr
+from pydantic import Field
 
+from duqtools.schema import BaseModel
 from duqtools.utils import formatter as f
 
-from ._basemodel import BaseModel
+from ..ids._schema import ImasBaseModel
+from ..schema import CoupledDim, Operation, OperationDim
+from ..schema.data_location import DataLocation
+from ..schema.matrix_samplers import (
+    CartesianProduct,
+    HaltonSampler,
+    LHSSampler,
+    SobolSampler,
+)
 
 
 def _default_sampler():
     from .matrix_samplers import CartesianProduct
     return CartesianProduct()
-
-
-def _default_system():
-    from ..systems.no_system import NoSystemModel
-    return NoSystemModel()
 
 
 class CreateConfigModel(BaseModel):
@@ -96,46 +100,3 @@ class CreateConfigModel(BaseModel):
         starting by the run number given by `run_out_start_at`.
 
         """))
-
-
-class ConfigModel(BaseModel):
-    """The options for the CLI are defined by this model."""
-    tag: str = Field(
-        '',
-        description=
-        'Create a tag for the runs to identify them in slurm or `data.csv`')
-
-    create: Optional[CreateConfigModel] = Field(
-        None,
-        description=
-        'Configuration for the create subcommand. See model for more info.')
-
-    extra_variables: Optional[VariableConfigModel] = Field(
-        None, description='Specify extra variables for this run.')
-
-    system: Union[NoSystemModel, Ets6SystemModel, JettoSystemModel] = Field(
-        default_factory=_default_system,
-        description='Options specific to the system used',
-        discriminator='name')
-
-    quiet: bool = Field(
-        False,
-        description=
-        'If true, do not output to stdout, except for mandatory prompts.')
-
-    _path: Union[Path, str] = PrivateAttr(None)
-
-
-from ..systems.ets import Ets6SystemModel  # noqa
-from ..systems.jetto import JettoSystemModel  # noqa
-from ..systems.no_system import NoSystemModel  # noqa
-from ._dimensions import CoupledDim, Operation, OperationDim  # noqa
-from ._imas import ImasBaseModel  # noqa
-from .data_location import DataLocation  # noqa
-from .matrix_samplers import (  # noqa
-    CartesianProduct, HaltonSampler, LHSSampler, SobolSampler,
-)
-from .variables import VariableConfigModel  # noqa
-
-CreateConfigModel.model_rebuild()
-ConfigModel.model_rebuild()
