+++ src/duqtools/ids/_handle.py
@@ -12,11 +12,11 @@ from pydantic import field_validator
 
 from ..config import lookup_vars, var_lookup
 from ..operations import add_to_op_queue
-from ..schema import ImasBaseModel
 from ._copy import copy_ids_entry
 from ._imas import imas, imasdef
 from ._mapping import IDSMapping
 from ._rebase import squash_placeholders
+from ._schema import ImasBaseModel
 
 if TYPE_CHECKING:
     import xarray as xr
