+++ src/duqtools/ids/_schema.py
