+++ src/duqtools/schema/__init__.py
@@ -1,27 +1,33 @@
 from __future__ import annotations
 
 from ._basemodel import BaseModel, RootModel
-from ._dimensions import IDSOperation, IDSOperationDim, JettoOperation, OperationDim
-from ._imas import ImasBaseModel
-from ._jetto import JettoField, JettoVar, JsetField, NamelistField
+from ._dimensions import (
+    CoupledDim,
+    DimMixin,
+    IDSOperation,
+    IDSOperationDim,
+    IDSPathMixin,
+    Operation,
+    OperationDim,
+    OperatorMixin,
+)
 from ._ranges import ARange, LinSpace
-from ._variable import IDS2JettoVariableModel, IDSVariableModel, JettoVariableModel
+from .variables import IDSVariableModel
 
 __all__ = [
     'ARange',
     'BaseModel',
     'RootModel',
     'IDSOperation',
+    'OperatorMixin',
+    'DimMixin',
+    'OperationDim',
+    'CoupledDim',
+    'IDSPathMixin',
+    'IDSOperation',
     'IDSOperationDim',
+    'Operation',
     'IDSVariableModel',
-    'IDS2JettoVariableModel',
-    'ImasBaseModel',
-    'JettoField',
-    'JettoOperation',
-    'JettoVar',
-    'JettoVariableModel',
-    'JsetField',
     'LinSpace',
-    'NamelistField',
     'OperationDim',
 ]
