+++ src/duqtools/systems/jetto/__init__.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 from ._batchfile import write_batchfile
 from ._jettovar_to_json import jettovar_to_json
+from ._models import IDS2JettoVariableModel, JettoVariableModel
 from ._schema import JettoSystemModel
 from ._system import BaseJettoSystem, JettoSystem, JettoSystemV210921, JettoSystemV220922
 
@@ -12,6 +13,8 @@ __all__ = [
     'JettoSystemV210921',
     'JettoSystemV220922',
     'JettoSystemModel',
+    'JettoVariableModel',
+    'IDS2JettoVariableModel',
     'jettovar_to_json',
     'write_batchfile',
 ]
